Create Static.py
Animal Welfare Trust is on a visit to the circus camp to have a look at the four talking parrots added to the camp.
A parrot is identified by its name and color. Apart from this, the trust has asked to assign a unique number for each parrot. The unique number should begin with 7001 and should be auto-incremented by 1 for every new parrot added to the camp.

Identify the class name and attributes so as to represent parrots from the list given.

beak_color

__init__(name,color)

name

Parrot

counter->static

color

Parrots

unique_number

Write a Python program to implement the class chosen with its attributes and methods. Represent few parrots, display their names, color and unique number.

Note: Consider all the attributes to be private and methods to be public. Include getter methods for all the instance variables.